Fixed pointer error. functions return higher precision floating point.

+++ lib/iges_parser/include/fxt_utils.h
@@ -5,8 +5,8 @@
 #include <stdlib.h>
 
 int utils_to_int(char *str);
-float utils_to_float(char *str);
-float utils_replace_char(char* str, char find, char replace);
+double utils_to_float(char *str);
+double utils_replace_char(char* str, char find, char replace);
 int utils_to_array(char **ret, char *str, char *delim);
 
 #endif
